+++ src/prepdir/__init__.py
@@ -1,3 +1,6 @@
 """
 prepdir - Directory traversal utility to prepare project contents for review
 """
+from .main import run
+
+__version__ = "0.13.0"